user login:
- move the login class to mutations.py
- if the password is empty, then return wrong credential error message
